Application is now an iFrame app, and uses application_helper and the application controller to pass api variables between requests. currently app_redirect_to and app_link_to must be used, after we figure out all of the hotspots (where api keys and such must be passed) we should alias the methods to make passing transparent for future developers.
The Deluux stylesheet was adapted to make appearance uniform with the rest of the site. tiny_mce_image_list.js is a hack to allow for dynamic images to be listed in tinyMCE. Lots of other people are doing the same hackish thing, so we should be safe.

all of the "which friends are mentioned in this post" logic has been re-written to take place exclusively on the client side. Because this is an iframe app, we can no longer do AJAX haxx to the deluux site. Instead, I'm writing the JS arrays and datastructures using ruby code in the view layer (in app/views/notes/new right now, will abstract it out for edit capabilities) and periodically run the RegExs over the body of the post.